target/aarch64: fix "reset halt" on target without srst_nogate
The commit "aarch64: add support for \"reset halt\"" (6c0151623)
introducedthe ability to use Reset Catch to make CPU halts on reset
immediately.

Although it's useful, it requires that JTAG logic will not be gated by
pulling nSRST low, which is the option set by srst_nogate.

For targets do not support srst_nogate, this commit breaks the original
"reset halt" logic used to be working.

This patch fixed the issue by limiting the use of Reset Catch only if
srst_nogate is set.

Signed-off-by: Weijie Gao <hackpascal@gmail.com>